pyproject.toml
[project]
name = "eveoniris-discord-bot"
version = "0.1.0"
description = "A discord bot for EveOniris"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "discord-py>=2.6.3",
    "sqlalchemy[asyncio]>=2.0.43",
]
